// backend/src/careers/enums/player-instruction.enum.ts
export enum PlayerInstruction {
  CARRY = 'CARRY',
  WEAK_SIDE = 'WEAK_SIDE',
  SPLIT_PUSH = 'SPLIT_PUSH',
  TEAMFIGHT = 'TEAMFIGHT',
  PLAY_FOR_TOP = 'PLAY_FOR_TOP',
  PLAY_FOR_MID = 'PLAY_FOR_MID',
  PLAY_FOR_BOT = 'PLAY_FOR_BOT',
  FARM_CARRY = 'FARM_CARRY',
  OBJECTIVE = 'OBJECTIVE',
  AGGRESSIVE_GANK = 'AGGRESSIVE_GANK',
  ROAM_TOP = 'ROAM_TOP',
  ROAM_BOT = 'ROAM_BOT',
  SUPPORT_JUNGLE = 'SUPPORT_JUNGLE',
  SCALING = 'SCALING',
  HYPER_CARRY = 'HYPER_CARRY',
  LANE_PRESSURE = 'LANE_PRESSURE',
  SAFE_FARM = 'SAFE_FARM',
  PROTECT_ADC = 'PROTECT_ADC',
  ROAM_MID = 'ROAM_MID',
  ROAM_UPPER = 'ROAM_UPPER',
  ENGAGE = 'ENGAGE',
  UTILITY = 'UTILITY',
}